Refactoring for audiobook site controllers
