+++ src/main/java/com/tsel/home/project/booklibrary/converter/AudioBookSiteConverter.java
@@ -8,6 +8,7 @@ public class AudioBookSiteConverter implements Converter<AudioBookSite, AudioBoo
     @Override
     public AudioBookSiteDTO convert(AudioBookSite entity) {
         return AudioBookSiteDTO.builder()
+            .id(entity.getId())
             .name(entity.getName())
             .build();
     }
