+++ src/main/java/com/tsel/home/project/booklibrary/utils/CollectionUtils.java
@@ -1,0 +1,16 @@
+package com.tsel.home.project.booklibrary.utils;
+
+import java.util.Collection;
+
+public final class CollectionUtils {
+
+    private CollectionUtils() {}
+
+    public static boolean isEmpty(Collection<?> collection) {
+        return collection == null || collection.isEmpty();
+    }
+
+    public static boolean isNotEmpty(Collection<?> collection) {
+        return !isEmpty(collection);
+    }
+}
