+++ src/main/java/com/tsel/home/project/booklibrary/utils/elements/ImageProvider.java
@@ -72,7 +72,7 @@ public final class ImageProvider {
 
     public Image resolveSmallCover(Book book) {
         if (isNotBlank(book.getCoverImgAbsolutePath())) {
-            Image cashedImage = cashedSmallImages.get(book.getKey());
+            Image cashedImage = cashedSmallImages.get(book.getId());
             if (cashedImage != null) {
                 return cashedImage;
             }
