+++ src/main/resources/com/tsel/home/project/booklibrary/view/audio-book-sites-connections-view.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onKeyPressed="#stageKeyPressed" prefHeight="675.0" prefWidth="484.0" style="-fx-background-color: #C7CCB3;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.tsel.home.project.booklibrary.controller.AudioBookSiteConnectionViewController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onKeyPressed="#stageKeyPressed" prefHeight="675.0" prefWidth="484.0" style="-fx-background-color: #C7CCB3;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.tsel.home.project.booklibrary.controller.impl.audiobook.AudioBookSiteConnectionViewController">
     <children>
         <AnchorPane layoutY="-1.0" prefHeight="125.0" prefWidth="484.0" style="-fx-background-color: #8B936B;">
             <children>
